The problem was that the $XDSPatientId was getting stepped on in the outbound fan-out.  So, if a request is to fan-out to 2.2 and 3.3, the two requests would go out, but both would have patientid^^^&3.3.1&ISO instead of one to 2.2 having patientid^^^&2.2.1&ISO and one to 3.3 with patientid^^^&3.3.1&ISO.  The result is that the response adapter would then just make two queries for 3.3, and you get two identical sets of 3.3 results (and no 2.2 results)........change the doc repo db so that all docs are only for 2.2 and none for 3.3, and you get no results because both queries run for 3.3 which now returns 0 results.
The problem was with my cloneRequest method, that makes a copy of the original AdhocQueryRequest for each fan-out request (this is required to prevent JAX concurrent modification exception).  In my cloneRequest, I created a new AdhocQueryRequest, but was not creating a new/cloned slotlist on the AdhocQuery........I just copied the current slotlist onto the new/cloned request.  The result was that all new/cloned requests had a reference to the same slotlist, so when the patientid is updated in the slotlist for each request (to replace the ^^^&aaid to the appropriate aaid for the request target), the result is that this new value is set for all fan-out requests, since each one has reference to same slotlist.

Current patch also contains some cleanup for the DQ aggregation processing, and a fix for url setting in DQ and PD.

Gateway-1382